+++ CHANGELOG.md
@@ -11,8 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Weapon classes (`carbine`, `polearm`, `sidearm`...).
 - Compendium Browser 'Classification' tab.
 - Mastery, High Mastery, and Grand Mastery proficiency will now set the default roll mode and add the additional rolls when appropriate.
-- Character Flag `Nimble Reflexes`.
-- When the system grants (dis)advantage on a roll, a tooltip will mention where the advantage is coming from.
+- Character Flags `Danger Sense`, `Heat Adaptation`, `Nimble Reflexes`, `Pathfinder`, `Sonic Sensitivity`, `Tree Climber`, `Two Livered`.
+- When the system grants (dis)advantage on a roll, a tooltip will mention where it is coming from.
 
 ### Changes